fix no skill  project

+++ src/main/java/com/grauman/amdocs/dao/ProjectsDAO.java
@@ -74,50 +74,51 @@ public class ProjectsDAO implements IProjectsDAO {
                         throw new SQLException("Project insertion failed.");
                 }
             }
+            if(!(newProject.getProductSkill().isEmpty() && newProject.getProductSkill().isEmpty())) {
 
-            StringBuilder insertProjectSkill = new StringBuilder("INSERT INTO projectskill (project_id, skill_id,skill_level)" +
-                    " VALUES (?, ?,?)");
-            int sizeSkillProduct = newProject.getProductSkill().size();
-            int sizeSkillTechnical = newProject.getTechnicalSkill().size();
-            for (int i = 0; i < (sizeSkillProduct + sizeSkillTechnical) - 1; i++) {
-                insertProjectSkill.append(", (?, ?, ?)");
-            }
-
-            try (PreparedStatement fetchInsertProjectSkill = connection.prepareStatement(String.valueOf(insertProjectSkill), Statement.RETURN_GENERATED_KEYS)) {
-                int counter = 0;
-                int i;
-                for (i = 1; i <= (sizeSkillProduct) * 3; i += 3) {
-                    fetchInsertProjectSkill.setInt(i, projectID);
-                    fetchInsertProjectSkill.setInt(i + 1, newProject.getProductSkill().get(counter).getId());
-                    fetchInsertProjectSkill.setInt(i + 2, newProject.getProductSkill().get(counter).getLevel());
-                    ++counter;
-                }
-                counter = 0;
-                for (; i <= (sizeSkillTechnical) * 3 + (sizeSkillProduct) * 3; i += 3) {
-                    fetchInsertProjectSkill.setInt(i, projectID);
-                    fetchInsertProjectSkill.setInt(i + 1, newProject.getTechnicalSkill().get(counter).getId());
-                    fetchInsertProjectSkill.setInt(i + 2, newProject.getTechnicalSkill().get(counter).getLevel());
-                    ++counter;
-
-                }
-
-                fetchInsertProjectSkill.executeUpdate();
-                try (ResultSet generatedID = fetchInsertProjectSkill.getGeneratedKeys()) {
-                    if (!generatedID.next()) {
-                        String deleteQueryProject = "DELETE FROM project WHERE id = ?";
-                        try (PreparedStatement fetchDeleteQueryProject = connection.prepareStatement(deleteQueryProject, Statement.RETURN_GENERATED_KEYS)) {
-                            fetchDeleteQueryProject.setInt(1, newProject.getId());
-                            fetchDeleteQueryProject.executeUpdate();
+                        StringBuilder insertProjectSkill = new StringBuilder("INSERT INTO projectskill (project_id, skill_id,skill_level)" +
+                                " VALUES (?, ?,?)");
+                        int sizeSkillProduct = newProject.getProductSkill().size();
+                        int sizeSkillTechnical = newProject.getTechnicalSkill().size();
+                        for (int i = 0; i < (sizeSkillProduct + sizeSkillTechnical) - 1; i++) {
+                            insertProjectSkill.append(", (?, ?, ?)");
                         }
 
-                        throw new SQLException("Skill Project insertion failed. Project Deleted");
-                    }
+                        try (PreparedStatement fetchInsertProjectSkill = connection.prepareStatement(String.valueOf(insertProjectSkill), Statement.RETURN_GENERATED_KEYS)) {
+                            int counter = 0;
+                            int i;
+                            for (i = 1; i <= (sizeSkillProduct) * 3; i += 3) {
+                                fetchInsertProjectSkill.setInt(i, projectID);
+                                fetchInsertProjectSkill.setInt(i + 1, newProject.getProductSkill().get(counter).getId());
+                                fetchInsertProjectSkill.setInt(i + 2, newProject.getProductSkill().get(counter).getLevel());
+                                ++counter;
+                            }
+                            counter = 0;
+                            for (; i <= (sizeSkillTechnical) * 3 + (sizeSkillProduct) * 3; i += 3) {
+                                fetchInsertProjectSkill.setInt(i, projectID);
+                                fetchInsertProjectSkill.setInt(i + 1, newProject.getTechnicalSkill().get(counter).getId());
+                                fetchInsertProjectSkill.setInt(i + 2, newProject.getTechnicalSkill().get(counter).getLevel());
+                                ++counter;
 
-                }
-            }
+                            }
 
+                            fetchInsertProjectSkill.executeUpdate();
+                            try (ResultSet generatedID = fetchInsertProjectSkill.getGeneratedKeys()) {
+                                if (!generatedID.next()) {
+                                    String deleteQueryProject = "DELETE FROM project WHERE id = ?";
+                                    try (PreparedStatement fetchDeleteQueryProject = connection.prepareStatement(deleteQueryProject, Statement.RETURN_GENERATED_KEYS)) {
+                                        fetchDeleteQueryProject.setInt(1, newProject.getId());
+                                        fetchDeleteQueryProject.executeUpdate();
+                                    }
 
-        }
+                                    throw new SQLException("Skill Project insertion failed. Project Deleted");
+                                }
+
+                            }
+                        }
+
+            }//if no skill on this project
+        }//close connection
 
         return newProject;
     }
